requirements.txt
flask-restplus==0.9.2
Flask-SQLAlchemy==2.1
Werkzeug==0.16.0
pyyaml
pandas
numpy
scikit-learn
scipy
sqlalchemy
